introduce-request-metrics-recorder-move: Hand per-request metrics over to RequestMetricsRecorder
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
collect_metrics + _request_has_grammar from TokenizerManager; paste into
RequestMetricsRecorder (drop @staticmethod, replace
``self: "RequestMetricsRecorder"`` → plain ``self``). Caller prefix
replacement: ``TokenizerManager.collect_metrics(self.request_metrics_recorder, ...)``
→ ``self.request_metrics_recorder.collect_metrics(...)``.

Refactor chain ID: introduce-request-metrics-recorder-move
